+++ mzLib/TestFlashLFQ/ChromatographicPeakTests.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TestFlashLFQ
+namespace Test
 {
     public class ChromatographicPeakTests
     {
